--- src/main/java/mate/academy/rickandmorty/exception/CharacterNotFoundException.java
@@ -1,7 +1,0 @@
-package mate.academy.rickandmorty.exception;
-
-public class CharacterNotFoundException extends RuntimeException {
-    public CharacterNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
