+++ src/main/java/mate/academy/rickandmorty/exception/HttpRequestException.java
@@ -1,0 +1,7 @@
+package mate.academy.rickandmorty.exception;
+
+public class HttpRequestException extends RuntimeException {
+    public HttpRequestException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
